Move generic iterator to it's own package
